+++ src/renderer/shared/api/storage/migration/index.ts
@@ -1,0 +1,2 @@
+export { migrateEvents } from './migration-1';
+export { migrateWallets } from './migration-2';
